Move the test to tests/
